# Reduce compilation time