Compilation takes a long time for NCDatasets on the first few function calls. This PR explores strategies to reduce the compile time for loading `NCDatasets` and `CFVariables`

- type stability of struct fields
- `precompile` functions
- `Dict{Symbol,Any}` instead of mixed type `NamedTuple`

Original:
```julia
julia>     using NCDatasets

julia>     filename = "/home/raf/.julia/dev/GeoData/test/data/tos_O1_2001-2002.nc"
"/home/raf/.julia/dev/GeoData/test/data/tos_O1_2001-2002.nc"

julia>     @time ds = NCDatasets.NCDataset(filename);
  0.827610 seconds (1.86 M allocations: 103.473 MiB, 8.01% gc time, 99.75% compilation time)

julia>     @time NCDatasets.variable(ds, "tos");
  0.000147 seconds (10 allocations: 880 bytes)

julia>     @time ds["tos"];
  0.570029 seconds (1.47 M allocations: 87.999 MiB, 26.39% gc time, 99.70% compilation time)
```
This PR:
```julia
julia>     using NCDatasets
[ Info: Precompiling NCDatasets [85f8d34a-cbdd-5861-8df4-14fed0d494ab]

julia>     filename = "/home/raf/.julia/dev/GeoData/test/data/tos_O1_2001-2002.nc"
"/home/raf/.julia/dev/GeoData/test/data/tos_O1_2001-2002.nc"

julia>     @time ds = NCDatasets.NCDataset(filename);
  0.626486 seconds (494.88 k allocations: 29.255 MiB, 12.36% gc time, 99.73% compilation time)

julia>     @time NCDatasets.variable(ds, "tos");
  0.000045 seconds (10 allocations: 880 bytes)

julia>     @time ds["tos"];
  0.134726 seconds (74.41 k allocations: 4.506 MiB, 99.38% compilation time)
```

It's especially faster (4x) for `getindex(::NCDataset, ::String)`. There are a lot of other optimizations that could be doe, but I'm opening this as is for now.