@Alexander-Barth 2021-08-06
Thanks for taking a look into precompile latency. The idea to use a named tuple for `storage_attrib` was that the compiler could generate fully type stable and specialized code when applying the scaling (scale_factor and add_offset) and the masking.
With your PR, would e.g. `fillvalue(v)` not become type unstable?