@rafaqz 2021-08-06
Yes there will be some Any fuelds. But its about a tradeoff... Up front compilatuon vs some type instability later, and only having to compile the expensive functions once. With the named tuple you have many different functions to compile, and getindex is much slower. Using a Dict seemed to help that a lot because there is nothing to infer.

But Dict/NamedTuple is probably not the most important thing here. The precompile helps a lot, and most structs in this package are currently not type stable, which adds a lot to compilation time and was relatively easy to fix for the main types. I didnt fix structs in e.g deffered.jl, but theyre also not stable.