@Alexander-Barth 2021-08-17
Can you amend your PR with keeping `storage_attrib`as a named tuple ? Maybe we can use also the `precompile` approach by using the most common combination of `fillvalue/scale_factor/add_offset/...` (but this could be a separate PR).