@rafaqz 2021-08-17
Sure it will be interesting to see the comparison separate from the other changes. I meant to do this earlier I've been a little busy lately.