@rafaqz 2021-08-19
With precompilation it looks like the `NamedTuple` is actually slightly faster. 

```julia
julia>     @time ds["tos"];
  0.094708 seconds (13.40 k allocations: 912.854 KiB, 98.85% compilation time)
```

I also fixed the type stability of multifile datasets and a few other minor things.