@rafaqz 2021-08-19
The test failures are some kind shared library errors on Julia 1.3, not a problem with this PR.