@Alexander-Barth 2021-08-20
I ran a git bisect with julia 1.3.1, and it seems that the issue arises first on commit 2671e4aeca9aa77915399fceb5292a094bf0e497 .