@Alexander-Barth 2021-08-20
Commenting these lines in `precompile.jl` makes 2671e4a  succeed in Julia 1.3.1:

        #precompile(getindex, (CF, Colon))                                                                                                                            
        #precompile(getindex, (CF, Int))                                                                                                                              
        #precompile(getindex, (CF, ntuple(Int, N)...))                                                                                                                
        #precompile(Array, (CF,))                                                                                                                                     
        #precompile(Array, (Var,))               

Maybe a bug in Julia 1.3 ? One approach could be to wrap `precompile.jl` in `if VERSION >= v"1.6" ... end`