@rafaqz 2021-08-20
Maybe compilation RAM usage or something. `>= v1.6` sounds like a good solution to me.